Fixed wrong type of `phase`
`phase` should be a `float` value that ranges between `0` and `1`.

+++ examples/audio/01-simple-playback/simple-playback.c
@@ -76,7 +76,7 @@ SDL_AppResult SDL_AppIterate(void *appstate)
         /* generate a 440Hz pure tone */
         for (i = 0; i < SDL_arraysize(samples); i++) {
             const int freq = 440;
-            const int phase = (total_samples_generated * freq) % 8000;
+            const float phase = (total_samples_generated * freq) % 8000;
             samples[i] = (float)SDL_sin(phase * 2 * SDL_PI_D / 8000.0);
             total_samples_generated++;
         }
